+++ src/extension.ts
@@ -151,7 +151,7 @@ export async function fixIssue(
             ionicState.selectedAndroidDeviceName = '';
           }
 
-          channelShow();
+          //channelShow();
           clearInterval(interval);
           finishCommand(tip);
           cancelObject.cancelled = true;
@@ -418,6 +418,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
   if (!ionicState.runWeb) {
     const summary = await reviewProject(rootPath, context, context.workspaceState.get('SelectedProject'));
+    ionicState.lastSummary = summary;
     if (summary?.project.isCapacitor) {
       showTips();
     }